Add support for building a shared library (libaptdec).   Exported functions and types prefixed with apt_   Callback added for getsample, instead of calling named function.   Allow library to be built without libpng or libsndfile. Add support for building with MSVC on Windows. Fix a few MSVC compiler warnings. Fix nonlinear radiance correction for NOAA19.

+++ src/filter.c
@@ -27,7 +27,7 @@ float fir(float *buff, const float *coeff, const int len) {
 	for (int i = 0; i < len; i++) {
 		r += buff[i] * coeff[i];
 	}
-	return r;
+	return (float)r;
 }
 
 /* IQ finite impulse response
@@ -61,5 +61,5 @@ float rsfir(double *buff, const float *coeff, const int len, const double offset
 		alpha = n - k;
 		out += buff[i] * (coeff[k] * (1.0 - alpha) + coeff[k + 1] * alpha);
 	}
-	return out;
+	return (float)out;
 }
